fix: 디렉토리 수정 및 채팅 페이지 헤더 삭제
