+++ app/api/chat/[chatId]/shares/route.ts
@@ -2,9 +2,9 @@ export const dynamic = 'force-dynamic';
 
 import { type NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 import { PrismaChatMessageRepository } from "@/infra/repositories/prisma/PrismaChatMessageListRepository";
-import { GetShareChatListUsecase } from "@/application/usecases/Chat/GetShareChatListUsecase";
+import { GetShareChatListUsecase } from "@/application/usecases/chat/GetShareChatListUsecase";
 
 export async function GET(
   req: NextRequest,
