+++ app/api/chat/[chatId]/together/route.ts
@@ -2,10 +2,10 @@ export const dynamic = 'force-dynamic';
 
 import { type NextRequest, NextResponse } from "next/server";
 import { PrismaGroupBuyChatMessageRepository } from "@/infra/repositories/prisma/PrismaGroupBuyChatMessageRepository";
-import { GetGroupBuyChatMessagesUsecase } from "@/application/usecases/Chat/GetGroupBuyChatMessagesUsecase";
-import { GetGroupBuyChatInfoUsecase } from "@/application/usecases/Chat/GetGroupBuyChatInfoUsecase";
+import { GetGroupBuyChatMessagesUsecase } from "@/application/usecases/chat/GetGroupBuyChatMessagesUsecase";
+import { GetGroupBuyChatInfoUsecase } from "@/application/usecases/chat/GetGroupBuyChatInfoUsecase";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/lib/authOptions";
 
 export async function GET(
   req: NextRequest,
