+++ application/usecases/Chat/GetGroupBuyChatInfoUsecase.ts
@@ -1,4 +1,4 @@
-import type { GroupBuyChatMessageRepository } from "@/domain/repositories/Chat/GroupBuyChatMessageRepository";
+import type { GroupBuyChatMessageRepository } from "@/domain/repositories/chat/GroupBuyChatMessageRepository";
 
 export class GetGroupBuyChatInfoUsecase {
   constructor(private readonly chatMessageRepository: GroupBuyChatMessageRepository) {}
