+++ domain/repositories/Chat/GroupBuyChatMessageRepository.ts
@@ -1,4 +1,4 @@
-import type { GroupBuyChatMessageListDto } from "@/application/usecases/Chat/dto/GroupBuyChatMessageListDto";
+import type { GroupBuyChatMessageListDto } from "@/application/usecases/chat/dto/GroupBuyChatMessageListDto";
 
 export interface GroupBuyChatMessageRepository {
   findMessagesByChatId(
